requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
instagram-private-api==1.6.0
python-dateutil==2.8.0
python-dotenv==0.10.3
redis==3.3.8
requests==2.22.0
six==1.12.0
urllib3==1.25.3
